requirements.txt
Flask==3.1.0
ollama==0.4.0
dropbox==12.0.2
pdfplumber==0.8.0
python-dotenv==1.0.0
